+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ## xel.rb 1.6.0  not yet released
 
+* stop using `_custom_functions` sub-context, go direct
+
 
 ## xel.rb 1.5.1  released 2024-07-19